feat: quadratic split feat: benchmark boilerplate

+++ app/main.hpp
@@ -2,21 +2,19 @@
 #define MAIN_HPP
 
 #include "gfx/gfx.hpp"
-#include "loader.hpp"
-#include "rtree.hpp"
 
-#include <iostream>
 #include <chrono>
+#include <fstream>
+#include <iostream>
 #include <vector>
-
-#include <SDL2/SDL_image.h>
 
 typedef std::chrono::high_resolution_clock Clock;
 
 constexpr double MIN_LON = -125.0, MAX_LON = -66.0;
 constexpr double MIN_LAT = 24.0, MAX_LAT = 50.0;
 
-inline std::vector<std::pair<double, double>> readBinaryFile(const std::string& filename) {
+inline std::vector<std::pair<double, double>>
+readBinaryFile(const std::string& filename) {
   std::ifstream infile(filename, std::ios::binary);
   if (!infile) {
     std::cerr << "Error: Could not open input file." << std::endl;
@@ -25,17 +23,19 @@ inline std::vector<std::pair<double, double>> readBinaryFile(const std::string& 
   size_t size;
   infile.read(reinterpret_cast<char*>(&size), sizeof(size));
   std::vector<std::pair<double, double>> coordinates(size);
-  infile.read(reinterpret_cast<char*>(coordinates.data()), size * sizeof(std::pair<double, double>));
+  infile.read(reinterpret_cast<char*>(coordinates.data()),
+              size * sizeof(std::pair<double, double>));
   infile.close();
   return coordinates;
 }
 
-inline std::pair<float, float> coord_to_pixel(const std::pair<double, double>& coord) {
+inline std::pair<double, double>
+coord_to_pixel(const std::pair<double, double>& coord) {
   const double xNorm = (coord.second - MIN_LON) / (MAX_LON - MIN_LON);
   const double yNorm = 1.0 - (coord.first - MIN_LAT) / (MAX_LAT - MIN_LAT);
 
-  float x = (xNorm * WIDTH);
-  float y = (yNorm * HEIGHT);
+  double x = (xNorm * WIDTH);
+  double y = (yNorm * HEIGHT);
 
   return {x, y};
 }
